// src/types/enums.ts
export enum ClassType {
  WARRIOR = 'warrior',
  ROGUE = 'rogue',
  MAGE = 'mage',
}

export enum SlotType {
  WEAPON = 'weapon',
  OFFHAND = 'offhand',
  HELMET = 'helmet',
  ARMOR = 'armor',
  GLOVES = 'gloves',
  BOOTS = 'boots',
  RING_LEFT = 'ringLeft',
  RING_RIGHT = 'ringRight',
  NECKLACE = 'necklace',
}

export enum Rarity {
  NORMAL = 'normal',
  MAGIC = 'magic',
  RARE = 'rare',
  LEGENDARY = 'legendary',
  ANCIENT = 'ancient',
}

export enum ScoreMode {
  BALANCED = 'balanced',
  CRIT = 'crit',
  ATK_SPEED = 'atkSpeed',
  TOUGHNESS = 'toughness',
  MAIN_ATTR = 'mainAttr',
}

export enum MonsterType {
  BALANCED = 'balanced',
  HIGH_HP = 'highHp',
  HIGH_ATK = 'highAtk',
  REWARD = 'reward',
}

export enum AffixType {
  SHARP = 'sharp',
  CRUEL = 'cruel',
  EAGLE_EYE = 'eagleEye',
  SWIFT = 'swift',
  FIRE = 'fire',
  ICE = 'ice',
  LIGHTNING = 'lightning',
  VITALITY = 'vitality',
  GUARDIAN = 'guardian',
  NIMBLE = 'nimble',
  FIRE_RES = 'fireRes',
  ICE_RES = 'iceRes',
  LIGHTNING_RES = 'lightningRes',
  GREED = 'greed',
  FORTUNE = 'fortune',
  EXPERIENCE = 'experience',
  LEECH = 'leech',
  MOVE_SPEED = 'moveSpeed',
}

export enum MonsterAffix {
  THORNS = 'thorns',
  BERSERK = 'berserk',
  SHIELD = 'shield',
  SPLIT = 'split',
  LEECH = 'leech',
  POISON = 'poison',
  FORTIFIED = 'fortified',
  SWIFT = 'swift',
}

export enum TalentBranch {
  WARRIOR_BERSERKER = 'warriorBerserker',
  WARRIOR_IRON = 'warriorIron',
  ROGUE_SHADOW = 'rogueShadow',
  ROGUE_WIND = 'rogueWind',
  MAGE_ELEMENTAL = 'mageElemental',
  MAGE_ARCANE = 'mageArcane',
}